+++ controllers/usernamespace/controller.go
@@ -15,7 +15,7 @@ package usernamespace
 import (
 	"context"
 
-	"github.com/eclipse-che/che-operator/pkg/deploy/certificates"
+	"github.com/eclipse-che/che-operator/pkg/deploy/tls"
 	"github.com/eclipse-che/che-operator/pkg/util"
 
 	"github.com/devfile/devworkspace-operator/pkg/constants"
@@ -129,7 +129,7 @@ func (r *CheUserNamespaceReconciler) commonRules(ctx context.Context, namesInChe
 }
 
 func (r *CheUserNamespaceReconciler) watchRulesForConfigMaps(ctx context.Context) handler.EventHandler {
-	rules := r.commonRules(ctx, certificates.CheAllCACertsConfigMapName)
+	rules := r.commonRules(ctx, tls.CheAllCACertsConfigMapName)
 	return handler.EnqueueRequestsFromMapFunc(
 		handler.MapFunc(func(obj client.Object) []reconcile.Request {
 			return asReconcileRequestsForNamespaces(obj, rules)
@@ -313,7 +313,7 @@ func (r *CheUserNamespaceReconciler) reconcileTrustedCerts(ctx context.Context, 
 	}
 
 	sourceMap := &corev1.ConfigMap{}
-	if err := r.client.Get(ctx, client.ObjectKey{Name: certificates.CheAllCACertsConfigMapName, Namespace: checluster.Namespace}, sourceMap); err != nil {
+	if err := r.client.Get(ctx, client.ObjectKey{Name: tls.CheAllCACertsConfigMapName, Namespace: checluster.Namespace}, sourceMap); err != nil {
 		if !errors.IsNotFound(err) {
 			return err
 		}
