+++ pkg/deploy/dashboard/deployment_dashboard.go
@@ -21,7 +21,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 
 	"github.com/eclipse-che/che-operator/pkg/deploy"
-	"github.com/eclipse-che/che-operator/pkg/deploy/certificates"
+	"github.com/eclipse-che/che-operator/pkg/deploy/tls"
 	"github.com/eclipse-che/che-operator/pkg/util"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -244,7 +244,7 @@ func (d *Dashboard) provisionCustomPublicCA(volumes []corev1.Volume, volumeMount
 		VolumeSource: corev1.VolumeSource{
 			ConfigMap: &corev1.ConfigMapVolumeSource{
 				LocalObjectReference: corev1.LocalObjectReference{
-					Name: certificates.CheAllCACertsConfigMapName,
+					Name: tls.CheAllCACertsConfigMapName,
 				},
 			},
 		},
