+++ pkg/deploy/tls/certificates_test.go
@@ -9,7 +9,7 @@
 // Contributors:
 //   Red Hat, Inc. - initial API and implementation
 //
-package certificates
+package tls
 
 import (
 	"context"
